Update combine_hires.F90
set_name subroutine, real*4 -> real*8 due to changes in the main program

+++ map/src/src_region/combine_hires.F90
@@ -561,7 +561,7 @@ print *, nlon, nlat
 ! ===============================================
       implicit none
 !
-      real            ::  lon, lat
+      real*8          ::  lon, lat
 
       character*1     ::  ew, sn
       character*2     ::  clat
